Move istio specific lua script to istio component
