+++ skaffold-operator.yaml
@@ -130,6 +130,7 @@ build:
             - pkg/component/networking/istio/charts/istio/istio-crds/crd-all.gen.yaml
             - pkg/component/networking/istio/charts/istio/istio-ingress
             - pkg/component/networking/istio/charts/istio/istio-istiod
+            - pkg/component/networking/istio/lua/add_kubernetes_namespace_name_to_record.lua
             - pkg/component/networking/nginxingress
             - pkg/component/networking/nodelocaldns/constants
             - pkg/component/networking/vpn/envoy
@@ -137,7 +138,6 @@ build:
             - pkg/component/networking/vpn/seedserver
             - pkg/component/observability/logging
             - pkg/component/observability/logging/fluentbit
-            - pkg/component/observability/logging/fluentbit/lua/add_kubernetes_namespace_name_to_record.lua
             - pkg/component/observability/logging/fluentbit/lua/add_tag_to_record.lua
             - pkg/component/observability/logging/fluentbit/lua/modify_severity.lua
             - pkg/component/observability/logging/fluentcustomresources
